code/requirements.txt
glfw
numpy
pyrr
PyOpenGL
OpenEXR